add dashboard stories unit tests
